refactor(texts): initial state context
