--- src/modules/texts/TextEdit/components/Markdown/constant/Markdown.constants.ts
@@ -1,9 +1,0 @@
-import { IDataText } from "modules/texts/TextEdit/types";
-
-export const INITIAL_STATE_DATA_TEXT: IDataText = {
-	title_board: "",
-	title_text: "",
-	text: "",
-	created_at: "",
-	public: false,
-};
