--- src/modules/texts/TextEdit/components/Markdown/constant/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./Markdown.constants";
-export * from "./Markdown.constants";
